Implement {: and fix (LOCAL)
This implements the Forth 2012 {: ... :} syntax for locals.  I
changed (LOCAL) to comply with the standard.  The standard assigns the
values in (LOCAL) in reverse order as pForth did.  Additionaly I
introduced (LAST-LOCAL) which does the assignments in the old order
which is useful to implement {.

* fth/ansilocs.fth (LV.MATCH): Return index as offset from end.
(LAST-LOCAL): Factored out from (LOCAL).
(LV.REVERSE, LV.SWAP, (LAST-LOCAL)): New helpers.

* fth/locals.fth:  Implement {:.
({): Use (LAST-LOCAL).

* fth/misc2.fth (ENVIRONMENT?): Define #LOCALS.

* fth/t_locals2.fth: Tests for {: and (LOCAL).

* build/unix/Makefile, build/mingw-crossbuild-linux/Makefile: Run test
  in t_locals2.fth.